# rest/rest/generated-sources/openapi/src/main/java/org/openapitools/api/DeviceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.DeviceDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-09-20T17:51:07.994232+02:00[Europe/Zurich]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "device-controller", description = "the device-controller API")
public interface DeviceApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * DELETE /device/{deviceName}
     *
     * @param deviceName  (required)
     * @return Not Found (status code 404)
     *         or OK (status code 200)
     */
    @Operation(
        operationId = "deleteDevice",
        tags = { "device-controller" },
        responses = {
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/device/{deviceName}",
        produces = { "*/*" }
    )
    
    default ResponseEntity<Void> deleteDevice(
        @Parameter(name = "deviceName", description = "", required = true, in = ParameterIn.PATH) @PathVariable("deviceName") String deviceName
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /device
     *
     * @return Not Found (status code 404)
     *         or OK (status code 200)
     */
    @Operation(
        operationId = "getAll1",
        tags = { "device-controller" },
        responses = {
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = DeviceDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/device",
        produces = { "*/*" }
    )
    
    default ResponseEntity<List<DeviceDto>> getAll1(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "[ { \"name\" : \"name\", \"eiXmlName\" : \"eiXmlName\", \"configurationValues\" : [ { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 }, { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 } ], \"status\" : \"status\" }, { \"name\" : \"name\", \"eiXmlName\" : \"eiXmlName\", \"configurationValues\" : [ { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 }, { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 } ], \"status\" : \"status\" } ]";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /device/{deviceName}
     *
     * @param deviceName  (required)
     * @return Not Found (status code 404)
     *         or OK (status code 200)
     */
    @Operation(
        operationId = "getDevice",
        tags = { "device-controller" },
        responses = {
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = DeviceDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/device/{deviceName}",
        produces = { "*/*" }
    )
    
    default ResponseEntity<DeviceDto> getDevice(
        @Parameter(name = "deviceName", description = "", required = true, in = ParameterIn.PATH) @PathVariable("deviceName") String deviceName
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"name\" : \"name\", \"eiXmlName\" : \"eiXmlName\", \"configurationValues\" : [ { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 }, { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /device
     *
     * @param deviceDto  (required)
     * @return Not Found (status code 404)
     *         or OK (status code 200)
     */
    @Operation(
        operationId = "insertOrUpdateDevice",
        tags = { "device-controller" },
        responses = {
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = DeviceDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/device",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<DeviceDto> insertOrUpdateDevice(
        @Parameter(name = "DeviceDto", description = "", required = true) @Valid @RequestBody DeviceDto deviceDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"name\" : \"name\", \"eiXmlName\" : \"eiXmlName\", \"configurationValues\" : [ { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 }, { \"val\" : \"val\", \"name\" : \"name\", \"id\" : 0 } ], \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}